app/requirements.txt
fastapi
uvicorn
openai
azure-search-documents
azure-cosmos
python-dotenv
httpx
jinja2
requests
